requirements.txt
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
asgiref==3.7.2
certifi==2023.11.17
cffi==1.16.0
Django==4.2.6
djangorestframework==3.14.0
install==1.3.5
minio==7.2.0
psycopg2-binary==2.9.9
pycparser==2.21
pycryptodome==3.19.0
pytz==2023.3.post1
sqlparse==0.4.4
typing_extensions==4.8.0
urllib3==2.1.0
